+++ MapboxCoreNavigation/RouteController.swift
@@ -173,7 +173,9 @@ open class RouteController: NSObject {
         
         rawLocation = locations.last
         
-        locations.forEach { navigator.updateLocation(for: MBFixLocation($0)) }
+        for location in locations {
+            navigator.updateLocation(for: MBFixLocation(location))
+        }
         
         let status = navigator.getStatusForTimestamp(location.timestamp)
         
@@ -281,11 +283,14 @@ open class RouteController: NSObject {
             delegate?.router?(self, didUpdate: progress, with: location, rawLocation: rawLocation)
             
             //Fire the notification (for now)
-            NotificationCenter.default.post(name: .routeControllerProgressDidChange, object: self, userInfo: [
-                RouteControllerNotificationUserInfoKey.routeProgressKey: progress,
-                RouteControllerNotificationUserInfoKey.locationKey: location, //guaranteed value
-                RouteControllerNotificationUserInfoKey.rawLocationKey: rawLocation //raw
-                ])
+            var userInfo: [RouteControllerNotificationUserInfoKey: Any] = [
+                .routeProgressKey: progress,
+                .locationKey: location, //guaranteed value
+                .rawLocationKey: rawLocation, //raw
+            ]
+            userInfo[.headingKey] = heading
+            userInfo[.rawHeadingKey] = heading // heading snapping not yet implemented (an unnecessary?)
+            NotificationCenter.default.post(name: .routeControllerProgressDidChange, object: self, userInfo: userInfo)
         }
     }
     
@@ -363,9 +368,11 @@ extension RouteController: Router {
         }
         
         delegate?.router?(self, willRerouteFrom: location)
-        NotificationCenter.default.post(name: .routeControllerWillReroute, object: self, userInfo: [
-            RouteControllerNotificationUserInfoKey.locationKey: location
-            ])
+        var userInfo: [RouteControllerNotificationUserInfoKey: Any] = [
+            .locationKey: location,
+        ]
+        userInfo[.headingKey] = heading
+        NotificationCenter.default.post(name: .routeControllerWillReroute, object: self, userInfo: userInfo)
         
         self.lastRerouteLocation = location